+++ src/core/async/modules/proxy/index.ts
@@ -22,6 +22,7 @@ import Super, {
 
 	BoundFn,
 
+	AsyncOptions,
 	AsyncCbOptions,
 	AsyncProxyOptions,
 
@@ -123,6 +124,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 
 	/**
 	 * Terminates the specified worker or a group of workers
+	 *
+	 * @alias
 	 * @param opts - options for the operation
 	 */
 	terminateWorker(opts: ClearProxyOptions<WorkerLikeP>): this;
@@ -146,7 +149,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Wraps the specified function.
+	 * Creates a new function that wraps the original and returns it.
+	 *
 	 * This method doesn't attach any hook or listeners to the object,
 	 * but if we cancel the operation by using one of Async's methods, like, `cancelProxy`,
 	 * the target function won't be invoked.
@@ -216,6 +220,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 
 	/**
 	 * Cancels the specified proxy function or a group of functions
+	 *
+	 * @alias
 	 * @param opts - options for the operation
 	 */
 	cancelProxy(opts: ClearProxyOptions<Function>): this;
@@ -315,13 +321,14 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Wraps the specified external request.
+	 * Creates a promise that wraps the passed request and returns it.
 	 *
 	 * This method doesn't attach any hook or listeners to the object,
 	 * but if we cancel the operation by using one of Async's methods, like, "cancelRequest",
 	 * the promise will be rejected.
 	 *
 	 * The request can be provided as a promise or function, that returns a promise.
+	 * Notice, the method uses `Async.promise`, but with a different namespace: `request` instead of `promise`.
 	 *
 	 * @param request
 	 * @param [opts] - additional options for the operation
@@ -337,7 +344,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Cancels the specified request
+	 * Cancels the specified request.
+	 * The canceled promise will be automatically rejected.
 	 *
 	 * @alias
 	 * @param [id] - link to the request (if not specified, then the operation will be applied for all registered tasks)
@@ -345,7 +353,10 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	cancelRequest(id?: Promise<unknown>): this;
 
 	/**
-	 * Cancels the specified request or a group of requests
+	 * Cancels the specified request or a group of requests.
+	 * The canceled promises will be automatically rejected.
+	 *
+	 * @alias
 	 * @param opts - options for the operation
 	 */
 	cancelRequest(opts: ClearOptionsId<Promise<unknown>>): this;
@@ -354,13 +365,17 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Cancels the specified request
+	 * Cancels the specified request.
+	 * The canceled promise will be automatically rejected.
+	 *
 	 * @param [id] - link to the request (if not specified, then the operation will be applied for all registered tasks)
 	 */
 	clearRequest(id?: Promise<unknown>): this;
 
 	/**
-	 * Cancels the specified request or a group of requests
+	 * Cancels the specified request or a group of requests.
+	 * The canceled promises will be automatically rejected.
+	 *
 	 * @param opts - options for the operation
 	 */
 	clearRequest(opts: ClearOptionsId<Promise<unknown>>): this;
@@ -369,13 +384,17 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Mutes the specified request
+	 * Mutes the specified request.
+	 * If the request is resolved during it muted, the promise wrapper will be rejected.
+	 *
 	 * @param [id] - link to the request (if not specified, then the operation will be applied for all registered tasks)
 	 */
 	muteRequest(id?: Promise<unknown>): this;
 
 	/**
-	 * Mutes the specified request or a group of requests
+	 * Mutes the specified request or a group of requests.
+	 * If the requests are resolved during muted, the promise wrappers will be rejected.
+	 *
 	 * @param opts - options for the operation
 	 */
 	muteRequest(opts: ClearOptionsId<Promise<unknown>>): this;
@@ -429,7 +448,220 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Wraps the specified promise.
+	 * Creates a new asynchronous iterable object from the specified iterable and returns it.
+	 * If the passed iterable doesn't have `Symbol.asyncIterator`, it will be created from a synchronous object iterator
+	 * (the synchronous iterator will also be preserved).
+	 *
+	 * Notice, until the created promise object isn't executed by invoking the `next` method,
+	 * any async operations won't be registered.
+	 *
+	 * @param iterable
+	 * @param [opts] - additional options for the operation
+	 *
+	 * @example
+	 * ```js
+	 * const async = new Async();
+	 *
+	 * for await (const el of async.iterable([1, 2, 3, 4])) {
+	 *   console.log(el);
+	 * }
+	 * ```
+	 */
+	iterable<T>(
+		iterable: Iterable<T> | AsyncIterable<T>,
+		opts: AsyncOptions
+	): AsyncIterable<T> | AsyncIterable<T> & Iterable<T> {
+		const
+			baseIterator = this.getBaseIterator(iterable);
+
+		if (baseIterator == null) {
+			return {
+				// @ts-ignore (type cast)
+				// eslint-disable-next-line require-yield
+				*[Symbol.asyncIterator]() {
+					return undefined;
+				}
+			};
+		}
+
+		let
+			globalError,
+			doneDelay = 0;
+
+		const newIterable = {
+			[Symbol.asyncIterator]: () => ({
+				[Symbol.asyncIterator]() {
+					return this;
+				},
+
+				next: () => {
+					if (globalError != null) {
+						return Promise.reject(globalError);
+					}
+
+					const promise = this.promise(Promise.resolve(baseIterator.next()), {
+						...opts,
+						name: this.namespaces.iterable,
+						onMutedResolve: (resolve, reject) => {
+							// Prevent an infinity loop if the iterable is already done
+							if (doneDelay > 0) {
+								setTimeout(() => {
+									resolve({value: undefined, done: true});
+								}, doneDelay);
+
+								return;
+							}
+
+							Promise.resolve(baseIterator.next()).then((res) => {
+								if (res.done) {
+									if (doneDelay === 0) {
+										doneDelay = 15;
+
+									} else if (doneDelay < 200) {
+										doneDelay *= 2;
+									}
+								}
+
+								resolve(res);
+							}, reject);
+						}
+					});
+
+					promise.catch((err) => {
+						if (Object.isDictionary(err) && err.type === 'clearAsync') {
+							globalError = err;
+						}
+					});
+
+					this.idsMap.set(newIterable, this.idsMap.get(promise) ?? promise);
+					return promise;
+				}
+			})
+		};
+
+		if (Object.isIterable(iterable[Symbol.iterator])) {
+			newIterable[Symbol.iterator] = iterable[Symbol.iterator];
+		}
+
+		return newIterable;
+	}
+
+	/**
+	 * Cancels the specified iterable object.
+	 * Notice that cancellation affects only objects that have already been activated by invoking the `next` method.
+	 * So, for example, canceled iterable will throw an error on the next invoking of `next`.
+	 *
+	 * @alias
+	 * @param [id] - link to the iterable (if not specified, then the operation will be applied for all registered tasks)
+	 */
+	cancelIterable(id?: AsyncIterable<unknown>): this;
+
+	/**
+	 * Cancels the specified iterable or a group of iterable.
+	 * Notice that cancellation affects only objects that have already been activated by invoking the `next` method.
+	 * So, for example, canceled iterable will throw an error on the next invoking of `next`.
+	 *
+	 * @alias
+	 * @param opts - options for the operation
+	 */
+	cancelIterable(opts: ClearOptionsId<AsyncIterable<unknown>>): this;
+	cancelIterable(task?: AsyncIterable<unknown> | ClearOptionsId<AsyncIterable<unknown>>): this {
+		return this.clearIterable(<any>task);
+	}
+
+	/**
+	 * Cancels the specified iterable object.
+	 * Notice that cancellation affects only objects that have already been activated by invoking the `next` method.
+	 * So, for example, canceled iterable will throw an error on the next invoking of `next`.
+	 *
+	 * @param [id] - link to the request (if not specified, then the operation will be applied for all registered tasks)
+	 */
+	clearIterable(id?: Promise<unknown>): this;
+
+	/**
+	 * Cancels the specified iterable object.
+	 * Notice that cancellation affects only objects that have already been activated by invoking the `next` method.
+	 * So, for example, canceled iterable will throw an error on the next invoking of `next`.
+	 *
+	 * @param opts - options for the operation
+	 */
+	clearIterable(opts: ClearOptionsId<Promise<unknown>>): this;
+	clearIterable(task?: Promise<unknown> | ClearOptionsId<Promise<unknown>>): this {
+		return this.cancelTask(task, this.namespaces.iterable);
+	}
+
+	/**
+	 * Mutes the specified iterable object.
+	 * Elements that are consumed during the object is muted will be ignored.
+	 * Notice that muting affects only objects that have already been activated by invoking the `next` method.
+	 *
+	 * @param [id] - link to the iterable (if not specified, then the operation will be applied for all registered tasks)
+	 */
+	muteIterable(id?: AsyncIterable<unknown>): this;
+
+	/**
+	 * Mutes the specified iterable object or a group of iterable objects.
+	 * Elements, that are consumed during the object is muted will be ignored.
+	 * Notice that muting affects only objects that have already been activated by invoking the `next` method.
+	 *
+	 * @param opts - options for the operation
+	 */
+	muteIterable(opts: ClearOptionsId<AsyncIterable<unknown>>): this;
+	muteIterable(task?: AsyncIterable<unknown> | ClearOptionsId<AsyncIterable<unknown>>): this {
+		return this.markTask('muted', task, this.namespaces.iterable);
+	}
+
+	/**
+	 * Unmutes the specified iterable object
+	 * @param [id] - link to the iterable (if not specified, then the operation will be applied for all registered tasks)
+	 */
+	unmuteIterable(id?: AsyncIterable<unknown>): this;
+
+	/**
+	 * Unmutes the specified iterable function or a group of iterable objects
+	 * @param opts - options for the operation
+	 */
+	unmuteIterable(opts: ClearOptionsId<AsyncIterable<unknown>>): this;
+	unmuteIterable(task?: AsyncIterable<unknown> | ClearOptionsId<AsyncIterable<unknown>>): this {
+		return this.markTask('!muted', task, this.namespaces.iterable);
+	}
+
+	/**
+	 * Suspends the specified iterable object.
+	 * Notice that suspending affects only objects that have already been activated by invoking the `next` method.
+	 *
+	 * @param [id] - link to the iterable (if not specified, then the operation will be applied for all registered tasks)
+	 */
+	suspendIterable(id?: AsyncIterable<unknown>): this;
+
+	/**
+	 * Suspends the specified iterable or a group of iterable objects.
+	 * Notice that suspending affects only objects that have already been activated by invoking the `next` method.
+	 *
+	 * @param opts - options for the operation
+	 */
+	suspendIterable(opts: ClearOptionsId<AsyncIterable<unknown>>): this;
+	suspendIterable(task?: AsyncIterable<unknown> | ClearOptionsId<AsyncIterable<unknown>>): this {
+		return this.markTask('paused', task, this.namespaces.iterable);
+	}
+
+	/**
+	 * Unsuspends the specified iterable object
+	 * @param [id] - link to the iterable (if not specified, then the operation will be applied for all registered tasks)
+	 */
+	unsuspendIterable(id?: AsyncIterable<unknown>): this;
+
+	/**
+	 * Unsuspends the specified iterable or a group of iterable objects
+	 * @param opts - options for the operation
+	 */
+	unsuspendIterable(opts: ClearOptionsId<AsyncIterable<unknown>>): this;
+	unsuspendIterable(task?: AsyncIterable<unknown> | ClearOptionsId<AsyncIterable<unknown>>): this {
+		return this.markTask('!paused', task, this.namespaces.iterable);
+	}
+
+	/**
+	 * Creates a new promise that wraps the passed promise and returns it.
 	 *
 	 * This method doesn't attach any hook or listeners to the object,
 	 * but if we cancel the operation by using one of Async's methods, like, "cancelPromise",
@@ -491,6 +723,25 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 				onMerge: (...args) => {
 					canceled = true;
 					return this.onPromiseMerge(resolve, reject)(...args);
+				},
+
+				onMutedCall: (link) => {
+					const
+						handlers = Array.concat([], opts?.onMutedResolve);
+
+					if (handlers.length > 0) {
+						for (let i = 0; i < handlers.length; i++) {
+							handlers[i].call(ctx, wrappedResolve, proxyReject);
+						}
+
+					} else {
+						reject({
+							...p,
+							link,
+							reason: 'muting',
+							type: 'clearAsync'
+						});
+					}
 				}
 			});
 
@@ -548,7 +799,8 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Cancels the specified promise
+	 * Cancels the specified promise.
+	 * The canceled promise will be automatically rejected.
 	 *
 	 * @alias
 	 * @param [id] - link to the promise (if not specified, then the operation will be applied for all registered tasks)
@@ -556,7 +808,10 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	cancelPromise(id?: Promise<unknown>): this;
 
 	/**
-	 * Cancels the specified promise or a group of promises
+	 * Cancels the specified promise or a group of promises.
+	 * The canceled promises will be automatically rejected.
+	 *
+	 * @alias
 	 * @param opts - options for the operation
 	 */
 	cancelPromise(opts: ClearProxyOptions<Promise<unknown>>): this;
@@ -565,13 +820,17 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Cancels the specified promise
+	 * Cancels the specified promise.
+	 * The canceled promise will be automatically rejected.
+	 *
 	 * @param [id] - link to the promise (if not specified, then the operation will be applied for all registered tasks)
 	 */
 	clearPromise(id?: Promise<unknown>): this;
 
 	/**
-	 * Cancels the specified promise or a group of promises
+	 * Cancels the specified promise or a group of promises.
+	 * The canceled promises will be automatically rejected.
+	 *
 	 * @param opts - options for the operation
 	 */
 	clearPromise(opts: ClearProxyOptions<Promise<unknown>>): this;
@@ -599,13 +858,17 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 	}
 
 	/**
-	 * Mutes the specified promise
+	 * Mutes the specified promise.
+	 * If the promise is resolved during it muted, the promise wrapper will be rejected.
+	 *
 	 * @param [id] - link to the promise (if not specified, then the operation will be applied for all registered tasks)
 	 */
 	mutePromise(id?: Promise<unknown>): this;
 
 	/**
-	 * Mutes the specified promise or a group of promises
+	 * Mutes the specified promise or a group of promises.
+	 * If the promises are resolved during muted, the promise wrappers will be rejected.
+	 *
 	 * @param opts - options for the operation
 	 */
 	mutePromise(opts: ClearOptionsId<Promise<unknown>>): this;
@@ -812,5 +1075,23 @@ export default class Async<CTX extends object = Async<any>> extends Super<CTX> {
 		}
 
 		return this;
+	}
+
+	/**
+	 * Returns an iterator from the passed iterable object.
+	 * Notice, an asynchronous iterator has more priority.
+	 *
+	 * @param [iterable]
+	 */
+	protected getBaseIterator<T>(
+		iterable: Iterable<T> | AsyncIterable<T>
+	): CanUndef<IterableIterator<T> | AsyncIterableIterator<T>> {
+		if (Object.isFunction(iterable[Symbol.asyncIterator])) {
+			return iterable[Symbol.asyncIterator]();
+		}
+
+		if (Object.isFunction(iterable[Symbol.iterator])) {
+			return iterable[Symbol.iterator]();
+		}
 	}
 }