+++ src/core/prelude/types/index.ts
@@ -105,6 +105,17 @@ extend(Object, 'isIterable', (value) => {
 	);
 });
 
+/** @see [[ObjectConstructor.isAsyncIterable]] */
+extend(Object, 'isAsyncIterable', (value) => {
+	if (value == null) {
+		return false;
+	}
+
+	return Boolean(
+		typeof Symbol === 'function' ? value[Symbol.asyncIterator] : typeof value['@@asyncIterator'] === 'function'
+	);
+});
+
 /** @see [[ObjectConstructor.isString]] */
 extend(Object, 'isString', (value) => typeof value === 'string');